# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Insurance Pricing API",
    "summary": "HTTP inference API for the insurance premium pricing model.",
    "description": "Production-style FastAPI surface for serving insurance pricing predictions.\n\n## Scope\n- score raw insurance business records\n- expose frequency, severity, and final premium predictions\n- keep the served model bundle pinned at startup through `INSURANCE_PRICING_RUN_ID`\n\n## Request Contract\nSend the raw columns from the historical scoring dataset, such as the ones present in `data/test.csv`. Feature engineering stays inside the service layer, so clients only need to send business-facing fields.\n\n## Response Headers\nEvery response includes `X-Request-ID` and `X-Process-Time-Ms` headers to help trace calls across clients, logs, and deployment environments.\n\n## Batch Behavior\nBatch endpoints preserve input order and return the optional `index` field when it is provided by the caller.\n\n## Product Quote Endpoints\nThe public `/quotes` endpoints persist end-user quote history scoped to an anonymous `X-Client-ID` header so the Next.js frontend can retrieve prior simulations.",
    "contact": {
      "name": "Insurance Pricing Team"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/admin/users": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List user accounts for administrators",
        "operationId": "admin_list_users",
        "parameters": [
          {
            "name": "X-Session-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/users/{user_id}": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Deactivate a user account",
        "operationId": "admin_deactivate_user",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          },
          {
            "name": "X-Session-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "User not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/quotes": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List recent quotes for administrators",
        "operationId": "admin_list_quotes",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "X-Session-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminQuoteListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/quotes/{quote_id}": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Delete a persisted quote",
        "operationId": "admin_delete_quote",
        "parameters": [
          {
            "name": "quote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quote Id"
            }
          },
          {
            "name": "X-Session-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Quote not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AdminQuoteListResponse": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count",
            "description": "Number of quotes returned to the admin."
          },
          "quotes": {
            "items": {
              "$ref": "#/components/schemas/AdminQuoteSummaryResponse"
            },
            "type": "array",
            "title": "Quotes",
            "description": "Recent quotes available to the admin console."
          }
        },
        "type": "object",
        "required": [
          "count",
          "quotes"
        ],
        "title": "AdminQuoteListResponse",
        "examples": [
          {
            "count": 1,
            "quotes": [
              {
                "created_at_utc": "2026-03-23T10:15:00+00:00",
                "id": "30bc53c8-5112-4a49-a204-5e8de5e7dcc5",
                "marque_vehicule": "RENAULT",
                "modele_vehicule": "CLIO",
                "owner_email": "client@nova-assurances.fr",
                "prime_prediction": 201.25504,
                "run_id": "base_v2_catboost_two_part_tweedie_1.3_train_smoke_42_classic_none_none",
                "type_contrat": "Maxi",
                "user_id": "7416b54b-26a0-4e79-9c27-aea0101f8d55"
              }
            ]
          }
        ]
      },
      "AdminQuoteSummaryResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Identifier of the persisted quote."
          },
          "created_at_utc": {
            "type": "string",
            "format": "date-time",
            "title": "Created At Utc",
            "description": "UTC timestamp at which the quote was created."
          },
          "run_id": {
            "type": "string",
            "title": "Run Id",
            "description": "Model bundle identifier used to compute the quote."
          },
          "type_contrat": {
            "type": "string",
            "title": "Type Contrat",
            "description": "Insurance contract type."
          },
          "marque_vehicule": {
            "type": "string",
            "title": "Marque Vehicule",
            "description": "Vehicle brand."
          },
          "modele_vehicule": {
            "type": "string",
            "title": "Modele Vehicule",
            "description": "Vehicle model."
          },
          "prime_prediction": {
            "type": "number",
            "title": "Prime Prediction",
            "description": "Final premium prediction for the quote."
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id",
            "description": "Owner user identifier when attached."
          },
          "owner_email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Email",
            "description": "Owner email when the quote is attached to an account."
          },
          "deleted_at_utc": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At Utc",
            "description": "Deletion timestamp when the quote has been removed by an admin."
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at_utc",
          "run_id",
          "type_contrat",
          "marque_vehicule",
          "modele_vehicule",
          "prime_prediction"
        ],
        "title": "AdminQuoteSummaryResponse",
        "examples": [
          {
            "created_at_utc": "2026-03-23T10:15:00+00:00",
            "id": "30bc53c8-5112-4a49-a204-5e8de5e7dcc5",
            "marque_vehicule": "RENAULT",
            "modele_vehicule": "CLIO",
            "owner_email": "client@nova-assurances.fr",
            "prime_prediction": 201.25504,
            "run_id": "base_v2_catboost_two_part_tweedie_1.3_train_smoke_42_classic_none_none",
            "type_contrat": "Maxi",
            "user_id": "7416b54b-26a0-4e79-9c27-aea0101f8d55"
          }
        ]
      },
      "AdminUserListResponse": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count",
            "description": "Number of users returned to the admin."
          },
          "users": {
            "items": {
              "$ref": "#/components/schemas/AdminUserSummaryResponse"
            },
            "type": "array",
            "title": "Users",
            "description": "User accounts visible to the admin."
          }
        },
        "type": "object",
        "required": [
          "count",
          "users"
        ],
        "title": "AdminUserListResponse",
        "examples": [
          {
            "count": 1,
            "users": [
              {
                "created_at_utc": "2026-03-23T13:00:00+00:00",
                "email": "client@nova-assurances.fr",
                "id": "7416b54b-26a0-4e79-9c27-aea0101f8d55",
                "is_active": true,
                "role": "customer"
              }
            ]
          }
        ]
      },
      "AdminUserSummaryResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Identifier of the user account."
          },
          "created_at_utc": {
            "type": "string",
            "format": "date-time",
            "title": "Created At Utc",
            "description": "UTC timestamp at which the user was created."
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Email address attached to the account."
          },
          "role": {
            "type": "string",
            "enum": [
              "customer",
              "admin"
            ],
            "title": "Role",
            "description": "Authorization role for the account."
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Whether the account is still active."
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at_utc",
          "email",
          "role",
          "is_active"
        ],
        "title": "AdminUserSummaryResponse",
        "examples": [
          {
            "created_at_utc": "2026-03-23T13:00:00+00:00",
            "email": "client@nova-assurances.fr",
            "id": "7416b54b-26a0-4e79-9c27-aea0101f8d55",
            "is_active": true,
            "role": "customer"
          }
        ]
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}